requirements.txt
asgiref==3.2.3
attrs==19.3.0
awsebcli==3.18.0
bcrypt==3.1.7
blessed==1.17.4
boto==2.49.0
boto3==1.12.9
botocore==1.15.9
cached-property==1.5.1
cement==2.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
confusable-homoglyphs==3.2.0
cryptography==2.9.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.3
django-allauth==0.41.0
django-cors-headers==3.2.1
django-crispy-forms==1.8.1
django-heroku==0.3.1
django-registration==3.1
django-rest-auth==0.9.5
django-ses==0.8.14
django-storages==1.9.1
django-webpack-loader==0.6.0
djangorestframework==3.11.0
docker==4.2.0
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
future==0.16.0
gunicorn==20.0.4
heroku==0.1.4
idna==2.7
importlib-metadata==1.6.0
jmespath==0.9.5
jsonschema==3.2.0
oauthlib==3.1.0
paramiko==2.7.1
pathspec==0.5.9
Pillow==7.0.0
postgres==3.0.0
psycopg2==2.7.5
psycopg2-binary==2.8.4
psycopg2-pool==1.1
pycparser==2.20
PyNaCl==1.3.0
pyrsistent==0.16.0
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
requests==2.20.1
requests-oauthlib==1.3.0
s3transfer==0.3.3
semantic-version==2.5.0
six==1.11.0
sqlparse==0.3.0
termcolor==1.1.0
texttable==1.6.2
urllib3==1.24.3
wcwidth==0.1.9
websocket-client==0.57.0
whitenoise==5.0.1
zipp==3.1.0
